requirements.txt
google-adk>=1.5.0
mcp>=1.10.1
google-genai>=1.24.0
discord.py>=2.5.2